scoliosis_ai/requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Vision Transformers
transformers>=4.30.0
timm>=0.9.0

# Quantum Computing
pennylane>=0.30.0
pennylane-lightning>=0.30.0

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0

# Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# LLM Integration
huggingface-hub>=0.16.0
accelerate>=0.20.0
bitsandbytes>=0.40.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0

# Medical Imaging
pydicom>=2.4.0
SimpleITK>=2.2.0
monai>=1.2.0

# Optimization
optuna>=3.2.0

# Statistical Analysis (PhD-level)
statsmodels>=0.14.0
pingouin>=0.5.3

# Experiment Tracking
mlflow>=2.3.0

# Paper Writing & Documentation
jupyterlab>=4.0.0
notebook>=7.0.0
